[Generic Serial] Review management
As generic "Serial" and "serialEvent" could be used by
HardwareSerial or USBSerial, defaut definition have to
be move away from uart to be properly wired.
