+++ cores/arduino/HardwareSerial.h
@@ -27,6 +27,7 @@
 #include <inttypes.h>
 
 #include "Stream.h"
+#include "uart.h"
 
 // Define constants and variables for buffering incoming serial data.  We're
 // using a ring buffer (I think), in which head is the index of the location
@@ -147,7 +148,5 @@ extern HardwareSerial Serial8;
 extern HardwareSerial Serial9;
 extern HardwareSerial Serial10;
 extern HardwareSerial SerialLP1;
-
-extern void serialEventRun(void) __attribute__((weak));
 
 #endif
